Orbit use periapsis instead of semi-major axis
A parabolic trajectory has an infinite semi-major axis; also, an hyperbolic
trajectory must have a negative semi-major axis. Thus, it is simpler to use the
periapsis, which exists in all situations and to derive the semi-major axis
from it.

A new alternative constructor is added, which allows to feed the semi-major
axis for circular or elliptic orbits and for hyperbolic trajectories.